refactor: redesign left side decoration panel
